# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwiftLingo - Telegram Mini App translation marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List translation jobs with filtering
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sourceLang
          in: query
          schema:
            type: string
        - name: targetLang
          in: query
          schema:
            type: string
        - name: specialization
          in: query
          schema:
            type: string
        - name: deliveryType
          in: query
          schema:
            type: string
        - name: budgetMin
          in: query
          schema:
            type: number
        - name: budgetMax
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: myJobs
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResult"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new translation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID with bids
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job (client only, own jobs)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job (client only, own open jobs)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /payments:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate payment for a contract (escrow capture)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{contractId}/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment for a contract (activates contract)
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment confirmed, contract activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user's notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, role, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, translator, admin]
        preferredLanguage:
          type: string
          enum: [en, uz, ru]
        isTranslatorApproved:
          type: boolean
        createdAt:
          type: string
...
    TranslatorProfile:
      type: object
      required: [id, userId, pricePerWord, sourceLanguages, targetLanguages, specializations, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["integer", "null"]
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        completedJobsCount:
          type: integer
        createdAt:
          type: string
...
    Job:
      type: object
      required: [id, clientId, title, sourceLang, targetLang, status, budgetMin, budgetMax, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        sourceLang:
          type: string
        targetLang:
          type: string
        wordCount:
          type: ["integer", "null"]
        budgetMin:
          type: number
        budgetMax:
          type: number
        deliveryType:
          type: string
          enum: [standard, express, urgent]
        specialization:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        bidsCount:
          type: integer
        createdAt:
          type: string
    JobListResult:
      type: object
      required: [jobs, total]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            client:
              $ref: "#/components/schemas/User"
            bids:
              type: array
              items:
                $ref: "#/components/schemas/Bid"
    JobInput:
      type: object
      required: [title, sourceLang, targetLang, budgetMin, budgetMax, deliveryType]
      properties:
        title:
          type: string
        description:
          type: string
        sourceLang:
          type: string
        targetLang:
          type: string
        wordCount:
          type: integer
        budgetMin:
          type: number
        budgetMax:
          type: number
        deliveryType:
          type: string
          enum: [standard, express, urgent]
        specialization:
          type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        wordCount:
          type: integer
        budgetMin:
          type: number
        budgetMax:
          type: number
        deliveryType:
          type: string
          enum: [standard, express, urgent]
        specialization:
          type: string
        status:
          type: string
          enum: [open, cancelled]
    Bid:
      type: object
      required: [id, jobId, translatorId, amount, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        translatorId:
          type: integer
        amount:
          type: number
        coverLetter:
          type: ["string", "null"]
        deliveryDays:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        translator:
          $ref: "#/components/schemas/TranslatorProfile"
        translatorUser:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, contractId, amount, status, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        amount:
          type: number
        platformFee:
          type: number
        translatorPayout:
          type: number
        status:
          type: string
          enum: [pending, confirmed, released, refunded]
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
    PaymentInput:
      type: object
      required: [contractId]
      properties:
        contractId:
          type: integer
        paymentMethod:
          type: string
...
    Notification:
      type: object
      required: [id, userId, type, title, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [new_bid, bid_accepted, payment_received, contract_activated, new_message, delivery_submitted, delivery_approved, application_submitted, application_approved, application_rejected]
        title:
          type: string
        body:
          type: ["string", "null"]
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
...