# lib/api-zod/src/generated/types/jobListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwiftLingo - Telegram Mini App translation marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface JobListResult {
  jobs: Job[];
  total: number;
}